+++ Source/Views/PolytempoNetwork/Polytempo_GraphicsViewRegion.h
@@ -54,7 +54,7 @@ public:
     void setImage(Image *img, var);
     void setText(String text);
     void setProgressbar(String text,
-                        int    time,
+                        float    time,
                         int    duration = 0);
     
     Polytempo_ViewContentType getContentType();
